refactor(@angular-devkit/architect): remove temporary index2 entrypoint
