+++ packages/angular_devkit/architect/builders/concat.ts
@@ -8,7 +8,7 @@
 import { json } from '@angular-devkit/core';
 import { from, of } from 'rxjs';
 import { concatMap, first, last, map, switchMap } from 'rxjs/operators';
-import { BuilderOutput, BuilderRun, createBuilder } from '../src/index2';
+import { BuilderOutput, BuilderRun, createBuilder } from '../src';
 import { Schema as OperatorSchema } from './operator-schema';
 
 export default createBuilder<json.JsonObject & OperatorSchema>((options, context) => {
