+++ packages/angular_devkit/architect/builders/true.ts
@@ -5,6 +5,6 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
-import { createBuilder } from '../src/index2';
+import { createBuilder } from '../src';
 
 export default createBuilder(() => ({ success: true }));
