--- packages/angular_devkit/architect/src/index2.ts
@@ -1,10 +1,0 @@
-/**
- * @license
- * Copyright Google Inc. All Rights Reserved.
- *
- * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://angular.io/license
- */
-export * from './api';
-export { Architect, ScheduleOptions } from './architect';
-export { createBuilder } from './create-builder';
